ml-reranker/requirements_cpu.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cpu
torchvision==0.16.2+cpu
torchaudio==2.1.2+cpu
asyncio
transformers[torch]
fastapi
uvicorn
python-dotenv
Pillow==9.5.0
sentence-transformers
pydantic